requirements.txt
Flask==3.0.3
anthropic>=0.40.0
python-dotenv>=1.0.0
numpy>=1.26.0
